Require the individual jquery-ui components
Instead of jquery-ui/themes/base/all. Internally jquery-ui uses @import
directives which Webpack could deal with, but which end up being
nonresolvable paths in the world of the asset pipeline and
asset-precompilation.

Closes #72

As a side note, we might want to trim this down to the ui components
we're actually using?