Prevent errors with datatables
Setting destroy: true as an option will check to see if a datatable
has already been initialized on a given DOM element; if so it will
destroy it and initialize it again. Solves the problem of
re-initialization we've been having for some-odd reason.

Give options in the DOM instead of within the JS as much as possible to
prevent having to conditionally give IDs or classes to objects and
initialize a separate table for each of them.