Add fixed headers and improved styling
For styling:
- datatables.net-bs/css/dataTables.bootstrap.css gets you nice big arrows
to show sort order.
- datatables.net-fixedheader-bs/css/fixedHeader.bootstrap.css allows
the fixedHeader plugin for datatables to work
- datatables.net-dt/css/jquery.dataTables.css styles datatables such
that the datatable will take up the entire width of the page instead of
being squished. It has more padding in between cells as well. It just
looks nicer, except for the arrows that the bootstrap css files get you.

For javascript:
- datatables.net/js/jquery.dataTables.js allows you to initialize
datatables like at all.
- datatables.net-fixedheader/js/dataTables.fixedHeader.js is the meat of
the fixedHeader plugin and, you kow, fixes the header.
- datatables.net-bs/js/dataTables.bootstrap.js is the least necessary of
this bunch but does give the search bar a class which allows nice
form-control bootstrap styling. It looks better so I'm keeping it.