allow logging in as pre-existing passenger, but not all of them?
I think it clutters the view, but maybe just allowing logging in as any passenger is fine, like whatever.
Of course, it bears thinking about that permanent passengers may attempt a login. Should they be allowed
to edit their information? Probably. I should account for all possibilities.

Using factory bot to create a user (and adjusting the factory so it's more in line with what factories 'should'
be according to the FactoryBot creators) has its perks, no need to clutter the sessions controller with methods
that create a passenger properly.

The application controller does allow for a non-persisted passenger to create themselves and I'm unable to do
this in development due to the absense of certain key/value pairs in the request environment. Oh well, you
can do it yourself while developing within the application controller if you need.