Just tell DataTables it's a date
Closes #188.

In looking into it, it seems like there are cases where DataTables
thinks the data source is something other than the DOM (?) A fix I found
was to explicity put `data-type="@data-sort"` on the header. But... if
we're going to do that, why not just use `data-type="date"`? Sorts like
a dream.