Passengers may have a verifying agency if permanent.
So we no longer want or need to hide the entire interface for entering a
verifying agency or expiration date if the passenger is permanent.
Indeed, the two sets of logic--hiding the interface on page load, but
merely disabling the expiration date--were at odds.

We only want to disable and remove input from the eligibility expiration
date if the passenger is marked as permanent. The expiration date was
already disabled for permanent passengers at page load.