Add custom validations with custom messages
Either a phone number OR an address is required, per ST Supervisors.
Add a before_save callback that wipes unnecessary data. A reason to wipe
the data is that the verifying agency of a passenger could change at
any time, and a situation could arise such as changing from one doctor
to disability services to a different doctor where old doctor's information
fills in the field, giving dispatchers more historical information than
is necessary and making it more likely that the field will simply be
filled with outdated data.