Two small fixes for the dispatch log (#501)
* Add form namespace for log entries

This changes how the ids for form elements are generated, and prevents
collisions between `for=` values. We had manually done that with the
pinned checkbox, but not the text box, leading to some accessibility
errors.

* Change log timestamp from link to span

It seemed wierd to me that the tamestamp was a "link" that didn't go
anywhere.