Switch from turbolinks to turbo (#488)
* Switch to turbo-rails gem

* Install and require turbo js

* Replace turbolinks events with turbo events

* Fix logout links

* data-confirm -> data-turbo-confirm

* Correct status on error

If a form submission re-renders the page they came from on error, Turbo
needs to see a non 200 code. Otherwise it's an error.

* Fix datatables/logout interaction

Turbolinks wasn't firing the event that tears-down Datatables on this
delete link. I think this would only be a problem in development, since
in prod, you get redirected elsewhere. But for the same reason, there's
no need to hold onto Turbolinks after logging out.